requirements.txt
fastapi
uvicorn
pydantic
mcp
firebase-admin
google-cloud-firestore
google-genai
python-dotenv
requests
beautifulsoup4
lxml
sympy
langgraph
langchain
langchain-google-genai
langchain-core
